Educational Game 2011
Some files ended up getting lost in my transition to online storage.
So this game is incomplete.

But, this game was created in 2011 by Travis Horne and a classmate as an
educational tool for the Bolivar Missouri middle school. The game had
animated characters that you had to dodge and when you made it to a new
level you had to answer a question to continue. Points were gained by
getting power ups and bonus when you answered questions. A high score
board was implemented but the handler for that appears to be one of the
missing files.